% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customvision_imgs.R
\name{browse_image}
\alias{browse_image}
\title{View an image uploaded to a Custom Vision project}
\usage{
browse_image(project, img_id, which = c("resized", "original",
  "thumbnail"), iteration = NULL)
}
\arguments{
\item{project}{A Custom Vision project.}

\item{img_id}{The ID of the image. You can use \code{\link{list_images}} to get the image IDs for this project.}

\item{which}{Which image to view: the resized version used for training (the default), the original uploaded image, or the thumbnail.}

\item{iteration}{The iteration ID (roughly, which model generation to use). Defaults to the latest iteration.}
}
\description{
View an image uploaded to a Custom Vision project
}
\details{
Images in a Custom Vision project are stored in Azure Storage. This function simply gets the URL for the uploaded image and displays it in your browser.
}
