% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision.R
\name{analyze}
\alias{analyze}
\alias{describe}
\alias{detect_objects}
\alias{detect_faces}
\alias{area_of_interest}
\alias{tag}
\alias{categorize}
\alias{read_text}
\alias{list_builtin_models}
\alias{make_thumbnail}
\title{Interface to Azure Computer Vision API}
\usage{
analyze(endpoint, image, domain = NULL, options = list(), ...)

describe(endpoint, image, ...)

detect_objects(endpoint, image, ...)

detect_faces(endpoint, image, ...)

area_of_interest(endpoint, image, ...)

tag(endpoint, image, language = "en", ...)

categorize(endpoint, image, ...)

read_text(endpoint, image, detect_orientation = TRUE, language = "en",
  ...)

list_builtin_models(endpoint)

make_thumbnail(endpoint, image, width, height, smart_crop = TRUE, ...,
  outfile = NULL)
}
\arguments{
\item{endpoint}{A computer vision endpoint.}

\item{image}{An image to be sent to the endpoint. This can be either a filename, a publicly accessible URL, or a raw vector holding the file contents.}

\item{domain}{For \code{analyze}, which builtin model to use to analyze the image.}

\item{options}{For \code{analyze}, a list of optional queries to obtain a broader analysis.}

\item{...}{Arguments passed to lower-level functions, and ultimately to \code{call_cognitive_endpoint}.}

\item{language}{For \code{tag} and \code{read_text}, a 2-character code indicating the language to use. The default is \code{en}, for English.}

\item{detect_orientation}{For \code{read_text}, whether to automatically determine the image's orientation.}

\item{width, height}{For \code{make_thumbnail}, The dimensions for the returned thumbnail.}

\item{smart_crop}{For \code{make_thumbnail}, whether to automatically determine the best location to crop for the thumbnail. Useful when the aspect ratio of the original image and the thumbnail don't match.}

\item{outfile}{For \code{make_thumbnail}, an optional filename for the generated thumbnail. If not provided, the thumbnail is returned as a raw vector.}
}
\description{
Interface to Azure Computer Vision API
}
