% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customvision_imgs.R
\name{add_images}
\alias{add_images}
\alias{list_images}
\alias{remove_images}
\title{Add, list and remove images for a project}
\usage{
add_images(project, images, tags = NULL, regions = NULL)

list_images(project, include = c("all", "tagged", "untagged"),
  as = c("ids", "dataframe", "list"), iteration = NULL)

remove_images(project, image_ids = list_images(project, "untagged", as =
  "ids"), confirm = TRUE)
}
\arguments{
\item{project}{A Custom Vision project.}

\item{images}{For \code{add_images}, the images to add (upload) to the project.}

\item{tags}{For \code{add_images}, optional tags to add to the images. Used for classification projects.}

\item{regions}{For \code{add_images}, an optional list of regions in the images that contain objects. Used for object detection projects.}

\item{include}{For \code{list_images}, which images to include in the list: untagged, tagged, or both (the default).}

\item{as}{For \code{list_images}, the return value: a vector of image IDs, a data frame of image metadata, or a list of metadata.}

\item{iteration}{For \code{list_images}, the iteration ID (roughly, which model generation to use). Defaults to the latest iteration.}

\item{image_ids}{For \code{remove_images}, the IDs of the images to remove from the project.}

\item{confirm}{For \code{remove_images}, whether to ask for confirmation first.}
}
\value{
For \code{add_images}, the vector of IDs of the uploaded images.

For \code{list_images}, based on the value of the \code{as} argument. The default is a vector of image IDs; \code{as="list"} returns a (nested) list of image metadata with one component per image; and \code{as="dataframe"} returns the same metadata but reshaped into a data frame.

For \code{remove_images}, NULL on successful removal.
}
\description{
Add, list and remove images for a project
}
\details{
The images to be uploaded can be specified as:
\itemize{
\item A vector of local filenames. All common image file formats are supported.
\item A vector of publicly accessible URLs.
\item A raw vector, or a list of raw vectors, holding the binary contents of the image files.
}

Uploaded images can be \emph{tagged} (for a classification project), which set the target labels for each image. To specify a single tag per image, set \code{tags} to be a vector. For multiple tags per image, \code{tags} should be a list, with each component being the vector of tags for the corresponding image. If \code{tags} has length 1, it is recycled to the length of \code{images} as a convenience feature for giving multiple images the same tag.

Specifying regions is used for an object detection project. The \code{regions} argument should be a nested list, with each component specifying the regions for the corresponding image.

Note that once uploaded, images are identified only by their ID; there is no general link back to the source filename or URL.
}
\seealso{
\code{\link{tag_uploaded_images}} to add tags to images if not done at upload time, \code{\link{untag_uploaded_images}}

\code{\link{add_tags}}, \code{\link{list_tags}}, \code{\link{remove_tags}}

\code{\link{customvision_project}}
}
