% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customvision_tags.R
\name{add_image_tags}
\alias{add_image_tags}
\alias{remove_image_tags}
\title{Tag and untag images uploaded to a project}
\usage{
add_image_tags(project, tags, image_ids = list_images(project,
  "untagged", as = "ids"))

remove_image_tags(project, image_ids = list_images(project, "tagged", as
  = "ids"), tags = list_tags(project, as = "ids"))
}
\arguments{
\item{project}{a Custom Vision project.}

\item{tags}{For \code{add_image_tags}, the tag labels to add to the images. For \code{remove_image_tags}, the tags (either text labels or IDs) to remove from images. The default for untagging is to remove all assigned tags.}

\item{image_ids}{The IDs of the images to tag or untag.}
}
\value{
The vector of IDs for the images affected, invisibly.
}
\description{
Tag and untag images uploaded to a project
}
\details{
\code{add_image_tags} is for tagging images that were uploaded previously, while \code{remove_image_tags} untags them. Adding tags does not remove previously assigned ones. Similarly, removing one tag from an image leaves any other tags intact.
}
\seealso{
\code{\link{add_images}}, \code{\link{list_tags}}
}
