% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{do_training_op}
\alias{do_training_op}
\alias{do_training_op.customvision_project}
\alias{do_prediction_op}
\alias{do_prediction_op.customvision_predictive_service}
\title{Carry out a Custom Vision operation}
\usage{
do_training_op(project, ...)

\method{do_training_op}{customvision_project}(project, op, ...)

do_prediction_op(service, ...)

\method{do_prediction_op}{customvision_predictive_service}(service, op, ...)
}
\arguments{
\item{project}{For \code{do_training_op}, a Custom Vision project.}

\item{op, ...}{Further arguments passed to \code{call_cognitive_endpoint}, and ultimately to the REST API.}

\item{service}{For \code{do_prediction_op}, a Custom Vision predictive service.}
}
\description{
Carry out a Custom Vision operation
}
\details{
These functions provide low-level access to the Custom Vision REST API. \code{do_training_op} is for working with the training endpoint, and \code{do_prediction_op} with the prediction endpoint. You can use them if the other tools in this package don't provide what you need.
}
\seealso{
\code{\link{customvision_training_endpoint}}, \code{\link{customvision_prediction_endpoint}},
\code{\link{customvision_project}}, \code{\link{customvision_predictive_service}}, \code{\link{call_cognitive_endpoint}}
}
