% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{computervision_endpoint}
\alias{computervision_endpoint}
\alias{customvision_training_endpoint}
\alias{customvision_prediction_endpoint}
\title{Endpoint objects for computer vision services}
\usage{
computervision_endpoint(url, ...)

customvision_training_endpoint(url, ...)

customvision_prediction_endpoint(url, ...)
}
\arguments{
\item{url}{The URL of the endpoint.}

\item{...}{Other arguments to pass to \code{\link[AzureCognitive:cognitive_endpoint]{AzureCognitive::cognitive_endpoint}}.}
}
\value{
An object inheriting from \code{cognitive_endpoint}. The subclass indicates the type of service/endpoint: Computer Vision, Custom Vision training, or Custom Vision prediction.
}
\description{
Endpoint objects for computer vision services
}
\details{
These are functions to create service-specific endpoint objects. They handle differences between the services in how they perform authentication, endpoint paths, and so on.
}
\seealso{
\code{\link{cognitive_endpoint}}, \code{\link{call_cognitive_endpoint}}
}
