% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVision.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cognitive_endpoint}
\alias{call_cognitive_endpoint}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AzureCognitive}{\code{\link[AzureCognitive]{call_cognitive_endpoint}}, \code{\link[AzureCognitive]{cognitive_endpoint}}}
}}

