% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customvision.R
\name{create_classification_project}
\alias{create_classification_project}
\alias{customvision_project}
\alias{create_object_detection_project}
\alias{create_customvision_project}
\alias{list_customvision_projects}
\alias{get_customvision_project}
\alias{update_customvision_project}
\alias{delete_customvision_project}
\title{Create, retrieve, update and delete Azure Custom Vision projects}
\usage{
create_classification_project(endpoint, name, domain = "general",
  export_target = c("none", "basic", "VAIDK"), multiple_tags = FALSE,
  description = NULL)

create_object_detection_project(endpoint, name, domain = "general",
  export_target = c("none", "basic", "VAIDK"), multiple_tags = FALSE,
  description = NULL)

create_customvision_project(endpoint, name, domain = "general",
  export_target = c("none", "basic", "VAIDK"), multiple_tags = FALSE,
  description = NULL, purpose = c("classification",
  "object_detection"))

list_customvision_projects(endpoint)

get_customvision_project(endpoint, name = NULL, id = NULL)

update_customvision_project(endpoint, name = NULL, id = NULL,
  domain = "general", export_target = c("none", "basic", "VAIDK"),
  multiple_tags = FALSE, description = NULL)

delete_customvision_project(object, ...)
}
\arguments{
\item{endpoint}{A custom vision endpoint.}

\item{name, id}{The name and ID of the project. At least one of these must be specified for \code{get_customvision_project}, \code{update_customvision_project} and \code{delete_customvision_project}. The name is required for \code{create_customvision_project} (the ID will be assigned automatically).}

\item{domain}{What kinds of images the model is meant to apply to. The default "general" means the model is suitable for use in a generic setting. Other, more specialised domains for classification include "food", "landmarks" and "retail"; for object detection the other possible domain is "logo".}

\item{export_target}{What formats are supported when exporting the model.}

\item{multiple_tags}{For classification models, Whether multiple categories (tags/labels) for an image are allowed. The default is \code{FALSE}, meaning an image represents one and only one category. Ignored for object detection models.}

\item{description}{An optional text description of the project.}

\item{purpose}{For \code{create_customvision_project}, what the model will be used for: either to classify objects in an image, or to detect whether an object is present in an image. Note that this setting cannot be changed once the project is created.}

\item{object}{For \code{delete_customvision_project}, either an endpoint, or a project object.}

\item{...}{Further arguments passed to lower-level methods.}
}
\value{
\code{delete_customvision_project} returns NULL invisibly, on a successful deletion. The others return an object of class \code{customvision_project}.
}
\description{
Create, retrieve, update and delete Azure Custom Vision projects
}
