% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customvision_training.R
\name{publish_model}
\alias{publish_model}
\alias{unpublish_model}
\title{Publish and unpublish a Custom Vision model iteration}
\usage{
publish_model(model, name, prediction_resource)

unpublish_model(model, confirm = TRUE)
}
\arguments{
\item{model}{A Custom Vision model iteration object.}

\item{name}{The name to assign to the published model.}

\item{prediction_resource}{The Azure prediction resource to publish to. This can either be a string containing the resource ID, or an AzureRMR resource object.}

\item{confirm}{For \code{unpublish_model}, whether to ask for confirmation first.}
}
\description{
Publish and unpublish a Custom Vision model iteration
}
\details{
Publishing a model makes it available to clients as a predictive service. Each iteration of the model can be published separately.
}
\seealso{
\code{\link{train_model}}, \code{\link{get_model}}, \code{\link{customvision_predictive_service}}, \code{\link{predict.classification_service}}, \code{\link{predict.object_detection_service}}
}
