% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customvision_imgs.R
\name{tag_uploaded_images}
\alias{tag_uploaded_images}
\alias{untag_uploaded_images}
\title{Tag and untag images uploaded to a project}
\usage{
tag_uploaded_images(project, tags, image_ids = list_images(project,
  "untagged", as = "ids"))

untag_uploaded_images(project, image_ids = list_images(project, "tagged",
  as = "ids"), tags = list_tags(project, as = "ids"))
}
\arguments{
\item{project}{a Custom Vision project.}

\item{tags}{For \code{tag_uploaded_images}, the tag labels to add to the images. For \code{untag_uploaded_images}, the tags (either text labels or IDs) to remove from images. The default for untagging is to remove all assigned tags.}

\item{image_ids}{The IDs of the images to tag or untag.}
}
\value{
The vector of IDs for the images affected, invisibly.
}
\description{
Tag and untag images uploaded to a project
}
\details{
\code{tag_uploaded_images} is for tagging images that were uploaded previously, while \code{untag_uploaded_images} untags them. Adding tags does not remove previously assigned ones. Similarly, removing one tag from an image leaves any other tags intact.
}
\seealso{
\code{\link{add_images}}, \code{\link{list_tags}}
}
